# docs/openapi.yaml
openapi: 3.0.3
info:
  title: express233-server API
  description: |
    游戏服务器集群推/拉一体交付控制平面 API。

    - **管理端**：Cookie 会话（`POST /api/login`）或 HTTP Basic Auth（初始化可用 `root/root`）
    - **节点端**：拉取 Token（下载使用 `token` 查询参数；Agent 心跳使用 `X-Express233-Token` 请求头）或 HTTP Basic Auth
  version: 1.0.0
servers:
  - url: http://localhost:23380
    description: 本地开发
...
paths:
...
  /api/projects/{id}/versions/{ver}/files:
    post:
      tags: [projects]
      summary: 上传单文件或 zip/tar/tar.gz/tgz 整包（自动解压）
      security: [{ cookieAuth: [] }, { basicAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: ver
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
                  description: 可选相对路径
                tags:
                  type: array
                  items: { type: string }
                  description: 可选文件标签；空或省略表示 *
      responses:
        "200":
          description: 已上传
  /api/projects/{id}/versions/{ver}/file-tags:
    get:
      tags: [projects]
      summary: 列出版本文件标签
      security: [{ cookieAuth: [] }, { basicAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: ver
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VersionFileTag"
    put:
      tags: [projects]
      summary: 设置单个版本文件标签
      security: [{ cookieAuth: [] }, { basicAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: ver
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileTagRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionFileTag"
    delete:
      tags: [projects]
      summary: 清空单个文件标签为 *
      security: [{ cookieAuth: [] }, { basicAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: ver
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 已清空
...
  /api/agent/guide/{topic}:
    get:
      tags: [agent]
      summary: 读取公开、脱敏的官方接入指南主题
      parameters:
        - { name: topic, in: path, required: true, schema: { type: string, enum: [overview, agent, gitea, github, security] } }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentGuideTopic"
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: express233_session
    basicAuth:
      type: http
      scheme: basic
    pullTokenHeader:
      type: apiKey
      in: header
      name: X-Express233-Token
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
...
    AgentGuideTopic:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        content: { type: string, description: 仅在单主题读取时返回；静态 Markdown 文本。 }
...
    VersionFileTag:
      type: object
      properties:
        path: { type: string }
        tags:
          type: array
          items: { type: string }
          example: ["linux-amd64", "linux"]
    FileTagRequest:
      type: object
      required: [path]
      properties:
        path: { type: string }
        tags:
          type: array
          items: { type: string }
          description: 空数组表示 *
...